requirements.txt
fastapi
uvicorn
tortoise-orm[aiomysql]
python-jose
dotenv
